refactor(core): organize observability modules
