+++ packages/core/package.json
@@ -18,6 +18,7 @@
   },
   "exports": {
     "./public": "./src/public/index.ts",
+    "./observability": "./src/observability/index.ts",
     "./session/runner": "./src/session/runner/index.ts",
     "./system-context": "./src/system-context/index.ts",
     "./*": "./src/*.ts"
